thermocouple libs split and generalized
